Cleaned up main file, moved UI to another file. Cleaning up still WIP
